run_many.m
This is the final wrapper for the entire project. This program will automatically increase the noise level from 0 to 0.5 by an amount of 0.05 at each time. At each noise level, it will apply our spectral clustering algorithm, attempting to reconstruct the original graph and compare the similarity between the result and the original graph through computing its Jaccard Index. At the very last point, it will plot a graph with respect to noise level and Jaccard Index to give us a intuitive idea about the effects of your spectral clustering algorithm.