IMPORTANT:please read README before running
This is the main program of this project. This program will first generate a specific clustering and a pure random noise graph based on the noise level. Then it will combine these two graphs to get the working graph. After this, this program will apply our spectral clustering algorithm on this working graph.At the very last, the program will compute the Jaccard Index between this resulting graph and the specific clustering.